+++ src/components/formFields/DoxenRadioDials.vue
@@ -68,15 +68,14 @@ import _cloneDeep from 'lodash.clonedeep';
 
 import {
   createRadioIdFor,
-  humanList,
   replaceWeirdCharacters,
-  wrappedHumanList
 } from '@/helpers/componentHelpers.js';
 import {
   createDisabledProp,
   createErrorMessageProp,
   createMessageProp,
   createModelValueProp,
+  createOptionsProp,
   label,
   required,
   styleTokens
@@ -95,6 +94,7 @@ const disabled = createDisabledProp('radio buttons');
 const errorMessage = createErrorMessageProp('radio buttons');
 const message = createMessageProp('radio buttons');
 const modelValue = createModelValueProp([String, Number, Boolean]);
+const options = createOptionsProp(COMPONENT_NAME);
 
 export default {
   name: COMPONENT_NAME,
@@ -114,51 +114,9 @@ export default {
     label,
     message,
     modelValue,
+    options,
     required,
-    styleTokens,
-    options: {
-      type: Array,
-      required: false,
-      default: function () {
-        return [];
-      },
-      description: 'Array of objects. Each object represents a radio button and contains a <code>name</code> (string) and a <code>value</code> (' + humanList(allowedOptionValueTypes) + ').',
-      example: [
-        '[',
-        '  {',
-        '    name: \'Kiva.org\',',
-        '    value: \'kiva\'',
-        '  },',
-        '  {',
-        '    name: \'Good.store\',',
-        '    value: \'goodstore\'',
-        '  }',
-        ']'
-      ].join('\n'),
-      validator: function (options) {
-        let valid = true;
-        if (Array.isArray(options)) {
-          options.forEach(function (option) {
-            if (
-              typeof(option) !== 'object' ||
-              Array.isArray(option) ||
-              !option.name ||
-              (
-                !allowedOptionValueTypes.includes(typeof(option.value)) &&
-                option.value !== null
-              )
-            ) {
-              console.warn('The ' + COMPONENT_NAME + ' options prop must be an array of objects with a name and a value that is a type of ' + wrappedHumanList(allowedOptionValueTypes));
-              console.warn('Example:\n<' + COMPONENT_NAME + ' :options="[{ name: \'Foo\', value: 2 }]" />');
-              valid = false;
-            }
-          });
-        } else {
-          valid = false;
-        }
-        return valid;
-      }
-    }
+    styleTokens
   },
   methods: {
     createRadioIdFor,
